requirements.txt
pydantic==1.10.9
email-validator==2.0.0.post2
Flask==2.3.2
SQLAlchemy==2.0.17
alembic[tz]==1.11.1
PyJWT==2.7.0
bcrypt==4.0.1
